Rename ElasticNetCCA's module file to match its class name
_sparsecca.py -> _elasticnetcca.py, matching the _gamcca.py/_gpcca.py/
_treecca.py convention and making it discoverable by name for users
already familiar with sklearn's ElasticNet.
