+++ cca_zoo/sparse/_elasticnetcca.py
